Split the models into separate files for ease of extension.
