+++ seirsplus/models/__init__.py
@@ -1,0 +1,5 @@
+"""SEIRS and extended SEIRS models."""
+
+from .extended_seirs_network_model import ExtSEIRSNetworkModel
+from .seirs_model import SEIRSModel
+from .seirs_network_model import SEIRSNetworkModel
